backend/requirements.txt
Django==4.2.17
djangorestframework==3.14.0
django-cors-headers==4.1.0
gunicorn
psycopg2-binary
